Renamed context, avoiding a proper SE-style refactor cause I'm lazy
