+++ src/assembly/mod.rs
@@ -1,5 +1,5 @@
 pub mod explication;
-mod explication_context;
 mod explication_explicator;
 mod explication_util;
 pub mod parser;
+mod context;
